@@ -20,8 +20,8 @@
     }
   ],
   "require": {
-    "php": "^7.2.5|^8.0.2",
-    "laravel/framework": "^7.0|^8.0|^9.0"
+    "php": "^8.0.2",
+    "laravel/framework": "^9.0"
   },
   "autoload": {
     "psr-4": {
